[rename] psf file
